chore: moved automatic localizations loading outside of LocalizationManager
